代码格式化、添加pom文件

+++ study16/src/main/java/com/zslin/utils/ExcelTemplate.java
@@ -75,20 +75,23 @@ public class ExcelTemplate {
     /**
      * 存储某一方所对于的样式
      */
-    private Map<Integer,CellStyle> styles;
+    private Map<Integer, CellStyle> styles;
     /**
      * 序号的列
      */
     private int serColIndex;
-    private ExcelTemplate(){
+
+    private ExcelTemplate() {
 
     }
+
     public static ExcelTemplate getInstance() {
         return et;
     }
 
     /**
      * 从classpath路径下读取相应的模板文件
+     *
      * @param path
      * @return
      */
@@ -102,8 +105,10 @@ public class ExcelTemplate {
         }
         return this;
     }
+
     /**
      * 将文件写到相应的路径下
+     *
      * @param filepath
      */
     public void writeToFile(String filepath) {
@@ -116,17 +121,19 @@ public class ExcelTemplate {
             throw new RuntimeException("写入的文件不存在");
         } catch (IOException e) {
             e.printStackTrace();
-            throw new RuntimeException("写入数据失败:"+e.getMessage());
+            throw new RuntimeException("写入数据失败:" + e.getMessage());
         } finally {
             try {
-                if(fos!=null) fos.close();
+                if (fos != null) fos.close();
             } catch (IOException e) {
                 e.printStackTrace();
             }
         }
     }
+
     /**
      * 将文件写到某个输出流中
+     *
      * @param os
      */
     public void wirteToStream(OutputStream os) {
@@ -134,11 +141,13 @@ public class ExcelTemplate {
             wb.write(os);
         } catch (IOException e) {
             e.printStackTrace();
-            throw new RuntimeException("写入流失败:"+e.getMessage());
+            throw new RuntimeException("写入流失败:" + e.getMessage());
         }
     }
+
     /**
      * 从某个路径来读取模板
+     *
      * @param path
      * @return
      */
@@ -155,6 +164,7 @@ public class ExcelTemplate {
 
     /**
      * 创建相应的元素，基于String类型
+     *
      * @param value
      */
     public void createCell(String value) {
@@ -163,24 +173,28 @@ public class ExcelTemplate {
         c.setCellValue(value);
         curColIndex++;
     }
+
     public void createCell(int value) {
         Cell c = curRow.createCell(curColIndex);
         setCellStyle(c);
-        c.setCellValue((int)value);
+        c.setCellValue(value);
         curColIndex++;
     }
+
     public void createCell(Date value) {
         Cell c = curRow.createCell(curColIndex);
         setCellStyle(c);
         c.setCellValue(value);
         curColIndex++;
     }
+
     public void createCell(double value) {
         Cell c = curRow.createCell(curColIndex);
         setCellStyle(c);
         c.setCellValue(value);
         curColIndex++;
     }
+
     public void createCell(boolean value) {
         Cell c = curRow.createCell(curColIndex);
         setCellStyle(c);
@@ -194,29 +208,33 @@ public class ExcelTemplate {
         c.setCellValue(value);
         curColIndex++;
     }
+
     public void createCell(BigInteger value) {
         Cell c = curRow.createCell(curColIndex);
         setCellStyle(c);
-        c.setCellValue(value==null?0:value.intValue());
+        c.setCellValue(value == null ? 0 : value.intValue());
         curColIndex++;
     }
+
     /**
      * 设置某个元素的样式
+     *
      * @param c
      */
     private void setCellStyle(Cell c) {
-        if(styles.containsKey(curColIndex)) {
+        if (styles.containsKey(curColIndex)) {
             c.setCellStyle(styles.get(curColIndex));
         } else {
             c.setCellStyle(defaultStyle);
         }
     }
+
     /**
      * 创建新行，在使用时只要添加完一行，需要调用该方法创建
      */
     public void createNewRow() {
-        if(lastRowIndex>curRowIndex&&curRowIndex!=initRowIndex) {
-            sheet.shiftRows(curRowIndex, lastRowIndex, 1,true,true);
+        if (lastRowIndex > curRowIndex && curRowIndex != initRowIndex) {
+            sheet.shiftRows(curRowIndex, lastRowIndex, 1, true, true);
             lastRowIndex++;
         }
         curRow = sheet.createRow(curRowIndex);
@@ -232,43 +250,47 @@ public class ExcelTemplate {
         int index = 1;
         Row row = null;
         Cell c = null;
-        for(int i=initRowIndex;i<curRowIndex;i++) {
+        for (int i = initRowIndex; i < curRowIndex; i++) {
             row = sheet.getRow(i);
             c = row.createCell(serColIndex);
             setCellStyle(c);
             c.setCellValue(index++);
         }
     }
+
     /**
      * 根据map替换相应的常量，通过Map中的值来替换#开头的值
+     *
      * @param datas
      */
-    public void replaceFinalData(Map<String,String> datas) {
-        if(datas==null) return;
-        for(Row row:sheet) {
-            for(Cell c:row) {
+    public void replaceFinalData(Map<String, String> datas) {
+        if (datas == null) return;
+        for (Row row : sheet) {
+            for (Cell c : row) {
 //                if(c.getCellType()!=Cell.CELL_TYPE_STRING) continue;
                 String str = c.getStringCellValue().trim();
-                if(str.startsWith("#")) {
-                    if(datas.containsKey(str.substring(1))) {
+                if (str.startsWith("#")) {
+                    if (datas.containsKey(str.substring(1))) {
                         c.setCellValue(datas.get(str.substring(1)));
                     }
                 }
             }
         }
     }
+
     /**
      * 基于Properties的替换，依然也是替换#开始的
+     *
      * @param prop
      */
     public void replaceFinalData(Properties prop) {
-        if(prop==null) return;
-        for(Row row:sheet) {
-            for(Cell c:row) {
+        if (prop == null) return;
+        for (Row row : sheet) {
+            for (Cell c : row) {
 //                if(c.getCellType()!=Cell.CELL_TYPE_STRING) continue;
                 String str = c.getStringCellValue().trim();
-                if(str.startsWith("#")) {
-                    if(prop.containsKey(str.substring(1))) {
+                if (str.startsWith("#")) {
+                    if (prop.containsKey(str.substring(1))) {
                         c.setCellValue(prop.getProperty(str.substring(1)));
                     }
                 }
@@ -282,22 +304,23 @@ public class ExcelTemplate {
         lastRowIndex = sheet.getLastRowNum();
         curRow = sheet.createRow(curRowIndex);
     }
+
     /**
      * 初始化数据信息
      */
     private void initConfigData() {
         boolean findData = false;
         boolean findSer = false;
-        for(Row row:sheet) {
-            if(findData) break;
-            for(Cell c:row) {
+        for (Row row : sheet) {
+            if (findData) break;
+            for (Cell c : row) {
 //                if(c.getCellType()!=Cell.CELL_TYPE_STRING) continue;
                 String str = c.getStringCellValue().trim();
-                if(str.equals(SER_NUM)) {
+                if (str.equals(SER_NUM)) {
                     serColIndex = c.getColumnIndex();
                     findSer = true;
                 }
-                if(str.equals(DATA_LINE)) {
+                if (str.equals(DATA_LINE)) {
                     initColIndex = c.getColumnIndex();
                     initRowIndex = row.getRowNum();
                     curColIndex = initColIndex;
@@ -310,37 +333,39 @@ public class ExcelTemplate {
                 }
             }
         }
-        if(!findSer) {
+        if (!findSer) {
             initSer();
         }
     }
+
     /**
      * 初始化序号位置
      */
     private void initSer() {
-        for(Row row:sheet) {
-            for(Cell c:row) {
+        for (Row row : sheet) {
+            for (Cell c : row) {
 //                if(c.getCellType()!=Cell.CELL_TYPE_STRING) continue;
                 String str = c.getStringCellValue().trim();
-                if(str.equals(SER_NUM)) {
+                if (str.equals(SER_NUM)) {
                     serColIndex = c.getColumnIndex();
                 }
             }
         }
     }
+
     /**
      * 初始化样式信息
      */
     private void initStyles() {
         styles = new HashMap<Integer, CellStyle>();
-        for(Row row:sheet) {
-            for(Cell c:row) {
+        for (Row row : sheet) {
+            for (Cell c : row) {
 //                if(c.getCellType()!=Cell.CELL_TYPE_STRING) continue;
                 String str = c.getStringCellValue().trim();
-                if(str.equals(DEFAULT_STYLE)) {
+                if (str.equals(DEFAULT_STYLE)) {
                     defaultStyle = c.getCellStyle();
                 }
-                if(str.equals(STYLE)) {
+                if (str.equals(STYLE)) {
                     styles.put(c.getColumnIndex(), c.getCellStyle());
                 }
             }
